fork twitterstream:  generic http streaming - less twitterish, move json serialzation and consumption to handlers, add metrics
